PowerPC: Pass custom thread count to backend
This patch allows users to override the number of threads
used during model execution via the OLLAMA_NUM_THREADS environment variable.

If this variable is not set, an optimal thread count is set for Power
Arch only.

By default, the thread count is set to runtime.NumCPU(), which uses all available logical cores.
This can cause CPU thrashing and degraded performance.

In internal testing, tuning the thread count to an optimal value for the
`granite-3b:8b` model led to a **40% improvement in inference
throughput**, compared to the default thread setting.

Example usage:
  OLLAMA_NUM_THREADS=8 ollama run ...

Signed-off-by: Shalini Salomi Bodapati <Shalini.Salomi.Bodapati@ibm.com>